+++ CHANGELOG.md
@@ -22,3 +22,4 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Variables used as function parameters to not show error
 - fixed '' in If
 - fixed '@Get()' not showing error for undeclared variable
+- fixed ^Data.[].Name! and ^Class.Name! to show error when ! is missing